src/requirements.txt
flask
ddtrace
JSON-log-formatter
mysql-connector-python
kombu